[Library.API] - Modified Order columns data type to double. - Cleaned-up migrations. - Modified Order computed columns query.

+++ src/Services/Library/Library.API/DTOs/Lessons/LessonResult.cs
@@ -7,5 +7,5 @@ public record LessonResult
 	public int LessonId { get; set; }
 	public string Title { get; set; }
 	public string Description { get; set; }
-	public decimal Order { get; set; }
+	public double Order { get; set; }
 }
